basic component test

+++ libs/map/src/jade/ecs/transform.rs
@@ -1,6 +1,6 @@
 #[derive(Debug, Default, PartialEq, PartialOrd)]
 pub struct Transform
 {
-    pub pos: (f32, f32),
-    pub size: (f32, f32),
+    pub pos: (f64, f64),
+    pub size: (f64, f64),
 }
